Remove userlogs - this is not a good idea
