+++ pkg/tracing/context.go
@@ -9,6 +9,61 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
+// GatewayAttributes holds metadata specific to a gateway component, helpful for logging.
+type GatewayAttributes struct {
+	Namespace string
+	Plugin    string // Optional. Name for the gateway-plugin
+	Route     string // Endpoint of the gateway
+}
+
+// CloudEventBusAttributes holds metadata specific to the cloud event bus.
+type CloudEventBusAttributes struct {
+	Namespace string
+	EventID   string // Unique identifier for the event
+	Source    string // Source of the event
+	Subject   string // Subject of the event
+	EventType string // Type of the event
+}
+
+// InstanceMemoryAttributes holds common metadata for an instance, which is helpful for logging.
+type InstanceMemoryAttributes struct {
+	Namespace    string
+	InstanceID   string // Unique identifier for the instance
+	WorkflowPath string // Path of the workflow the instance belongs to
+	Callpath     string // Identifies the log-stream, legacy feature from the old engine
+	State        string // Memory state of the instance
+	Invoker      string
+	Status       core.LogStatus
+}
+
+// InstanceActionAttributes holds metadata for an instance action.
+type InstanceActionAttributes struct {
+	Namespace    string
+	InstanceID   string // Unique identifier for the instance
+	WorkflowPath string // Path of the workflow the instance belongs to
+	Callpath     string // Identifies the log-stream, legacy feature from the old engine
+	State        string // Memory state of the instance
+	ActionID     string // Unique identifier for the instance action
+	Invoker      string
+	Status       core.LogStatus
+}
+
+// InstanceAttributes holds common metadata for an instance, which is helpful for logging.
+type InstanceAttributes struct {
+	Namespace    string
+	InstanceID   string // Unique identifier for the instance
+	WorkflowPath string // Path of the workflow the instance belongs to
+	Callpath     string // Identifies the log-stream, legacy feature from the old engine
+	Invoker      string
+	Status       core.LogStatus
+}
+
+// SyncAttributes holds metadata specific to a Sync component, helpful for logging.
+type SyncAttributes struct {
+	Namespace string
+	SyncID    string // Unique identifier for the Sync
+}
+
 // LogContextKey defines the keys used for storing values in the context for structured logging.
 type LogContextKey int
